Address PR review: separate Linux-specific mounting code and tests to support cross-platform compilation
